`textparser_get_token_position/length()` should return size_t not int.

+++ include/textparser.h
@@ -109,8 +109,8 @@ EXPORT_TEXTPARSER const textparser_token_item *textparser_get_token_child(const 
 EXPORT_TEXTPARSER const textparser_token_item *textparser_get_token_next(const textparser_token_item *token);
 EXPORT_TEXTPARSER const char *textparser_get_token_type_str(const textparser_language_definition *language, const textparser_token_item *token);
 EXPORT_TEXTPARSER int textparser_get_token_type(const textparser_token_item *token);
-EXPORT_TEXTPARSER int textparser_get_token_position(const textparser_token_item *token);
-EXPORT_TEXTPARSER int textparser_get_token_length(const textparser_token_item *token);
+EXPORT_TEXTPARSER size_t textparser_get_token_position(const textparser_token_item *token);
+EXPORT_TEXTPARSER size_t textparser_get_token_length(const textparser_token_item *token);
 EXPORT_TEXTPARSER uint32_t textparser_get_token_text_color(const textparser_token_item *token);
 EXPORT_TEXTPARSER uint32_t textparser_get_token_text_background(const textparser_token_item *token);
 EXPORT_TEXTPARSER uint32_t textparser_get_token_text_flags(const textparser_token_item *token);
